Documentation and minor fixes
