+++ alkos/kernel/arch/x86_64/common-loader-all/CMakeLists.txt
@@ -1,4 +1,4 @@
-message(STATUS "Configuring common")
+message(STATUS "Configuring x86_64 common-loader-all")
 
 ############################## Finding Sources ###############################
