+++ alkos/kernel/arch/x86_64/common-loader-all/abi/terminal.hpp
