+++ alkos/kernel/arch/x86_64/common-loader-all/drivers/vga/vga.cpp
