+++ alkos/kernel/arch/x86_64/common-loader-all/loader_memory_manager/loader_memory_manager.tpp
